+++ src/django/api/constants.py
@@ -7,10 +7,6 @@ class CsvHeaderField:
     ADDRESS = 'address'
     LAT = 'lat'
     LNG = 'lng'
-    PPE_PRODUCT_TYPES = 'ppe_product_types'
-    PPE_CONTACT_PHONE = 'ppe_contact_phone'
-    PPE_CONTACT_EMAIL = 'ppe_contact_email'
-    PPE_WEBSITE = 'ppe_website'
 
 
 class ProcessingAction:
@@ -39,7 +35,6 @@ class FacilitiesQueryParams:
     COUNTRIES = 'countries'
     COMBINE_CONTRIBUTORS = 'combine_contributors'
     BOUNDARY = 'boundary'
-    PPE = 'ppe'
     EMBED = 'embed'
     PARENT_COMPANY = 'parent_company'
     FACILITY_TYPE = 'facility_type'